+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+
+## [0.3.8]
+
+- Most tools, such as vue-cli@3, configure babel-loader to ignore all files inside node_modules.
+- So `vue-easy-lightbox.common.js` is now a version compiled with babel.
+
 ## [0.3.7]
 
 - Fix package.json entry point error.